+++ src/test/scala/edu/ie3/simona/util/ConfigUtilSpec.scala
@@ -14,23 +14,21 @@ import edu.ie3.datamodel.models.result.connector.{
   Transformer3WResult,
 }
 import edu.ie3.datamodel.models.result.system.LoadResult
-import edu.ie3.datamodel.models.result.{NodeResult, ResultEntity}
+import edu.ie3.datamodel.models.result.{
+  CongestionResult,
+  NodeResult,
+  ResultEntity,
+}
 import edu.ie3.simona.config.ConfigParams.ResultKafkaParams
-import edu.ie3.simona.config.RuntimeConfig._
 import edu.ie3.simona.config.OutputConfig.GridOutputConfig
-import edu.ie3.simona.config.{OutputConfig, SimonaConfig}
+import edu.ie3.simona.config.RuntimeConfig._
 import edu.ie3.simona.config.SimonaConfig.{apply => _, _}
+import edu.ie3.simona.config.{OutputConfig, SimonaConfig}
 import edu.ie3.simona.event.notifier.NotifierConfig
 import edu.ie3.simona.exceptions.InvalidConfigParameterException
 import edu.ie3.simona.test.common.{ConfigTestData, UnitSpec}
 import edu.ie3.simona.util.ConfigUtil.NotifierIdentifier._
-import edu.ie3.simona.util.ConfigUtil.{
-  EmConfigUtil,
-  GridOutputConfigUtil,
-  NotifierIdentifier,
-  OutputConfigUtil,
-  ParticipantConfigUtil,
-}
+import edu.ie3.simona.util.ConfigUtil._
 import org.scalatest.prop.{TableDrivenPropertyChecks, TableFor2}
 
 import java.util.UUID
@@ -643,32 +641,51 @@ class ConfigUtilSpec
         Table(
           ("config", "expected"),
           (
-            new GridOutputConfig(false, false, false, false, false),
+            new GridOutputConfig(
+              false, false, false, false, false, false,
+            ),
             Set.empty[Class[_ <: ResultEntity]],
           ),
           (
-            new GridOutputConfig(true, false, false, false, false),
+            new GridOutputConfig(
+              true, false, false, false, false, false,
+            ),
+            Set(classOf[CongestionResult]),
+          ),
+          (
+            new GridOutputConfig(
+              false, true, false, false, false, false,
+            ),
             Set(classOf[LineResult]),
           ),
           (
-            new GridOutputConfig(false, true, false, false, false),
+            new GridOutputConfig(
+              false, false, true, false, false, false,
+            ),
             Set(classOf[NodeResult]),
           ),
           (
-            new GridOutputConfig(false, false, true, false, false),
+            new GridOutputConfig(
+              false, false, false, true, false, false,
+            ),
             Set(classOf[SwitchResult]),
           ),
           (
-            new GridOutputConfig(false, false, false, true, false),
+            new GridOutputConfig(
+              false, false, false, false, true, false,
+            ),
             Set(classOf[Transformer2WResult]),
           ),
           (
-            new GridOutputConfig(false, false, false, false, true),
+            new GridOutputConfig(
+              false, false, false, false, false, true,
+            ),
             Set(classOf[Transformer3WResult]),
           ),
           (
-            new GridOutputConfig(true, true, true, true, true),
+            new GridOutputConfig(true, true, true, true, true, true),
             Set(
+              classOf[CongestionResult],
               classOf[LineResult],
               classOf[NodeResult],
               classOf[SwitchResult],