+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Added congestion detection [#1186](https://github.com/ie3-institute/simona/issues/1186)
+
+
 ## [4.0.0] - 2025-05-09
 
 ### Added